requirements.txt
python-telegram-bot==13.14
requests==2.28.1